:recycle: (Mailing) Move mail sending to external sendMail function
